Split Python package into multiple modules
